+++ index.js
@@ -20,8 +20,10 @@ module.exports = {
     'plugin:jest/recommended',
     'plugin:jest-formatting/strict',
     'plugin:react-hooks/recommended',
+    'plugin:import/typescript',
+    'plugin:typescript-enum/recommended',
   ],
-  plugins: ['backpack', 'prettier', 'jest-formatting'],
+  plugins: ['backpack', 'prettier', 'jest-formatting', 'typescript-enum'],
   rules: {
     'prettier/prettier': 'error',
     'valid-jsdoc': ['error'],
@@ -102,6 +104,21 @@ module.exports = {
       },
     ],
 
+    // Ensure consistent use of file extension within the import path
+    // The airbnb config we extend does not support TypeScript so we override it here
+    // https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/extensions.md
+    'import/extensions': [
+      'error',
+      'ignorePackages',
+      {
+        js: 'never',
+        jsx: 'never',
+        mjs: 'never',
+        ts: 'never',
+        tsx: 'never',
+      },
+    ],
+
     // Additional ESLint rules for directive comments of ESLint
     // See: https://github.com/mysticatea/eslint-plugin-eslint-comments
     'eslint-comments/disable-enable-pair': ['warn', { allowWholeFile: true }],
@@ -124,4 +141,68 @@ module.exports = {
 
     'jest/no-large-snapshots': 'off',
   },
+  overrides: [
+    {
+      files: ['**/*.ts?(x)'],
+      parser: '@typescript-eslint/parser',
+      parserOptions: {
+        ecmaVersion: 2018,
+        sourceType: 'module',
+        ecmaFeatures: {
+          jsx: true,
+        },
+
+        // typescript-eslint specific options
+        warnOnUnsupportedTypeScriptVersion: true,
+      },
+      plugins: ['@typescript-eslint'],
+      // If adding a typescript-eslint version of an existing ESLint rule,
+      // make sure to disable the ESLint rule here.
+      rules: {
+        // TypeScript's `noFallthroughCasesInSwitch` option is more robust (#6906)
+        'default-case': 'off',
+        // 'tsc' already handles this (https://github.com/typescript-eslint/typescript-eslint/issues/291)
+        'no-dupe-class-members': 'off',
+        // 'tsc' already handles this (https://github.com/typescript-eslint/typescript-eslint/issues/477)
+        'no-undef': 'off',
+
+        // Add TypeScript specific rules (and turn off ESLint equivalents)
+        '@typescript-eslint/consistent-type-assertions': 'warn',
+        'no-array-constructor': 'off',
+        '@typescript-eslint/no-array-constructor': 'warn',
+        'no-redeclare': 'off',
+        '@typescript-eslint/no-redeclare': 'warn',
+        'no-use-before-define': 'off',
+        '@typescript-eslint/no-use-before-define': [
+          'warn',
+          {
+            functions: false,
+            classes: false,
+            variables: false,
+            typedefs: false,
+          },
+        ],
+        'no-unused-expressions': 'off',
+        '@typescript-eslint/no-unused-expressions': [
+          'error',
+          {
+            allowShortCircuit: true,
+            allowTernary: true,
+            allowTaggedTemplates: true,
+          },
+        ],
+        'no-unused-vars': 'off',
+        '@typescript-eslint/no-unused-vars': [
+          'warn',
+          {
+            args: 'none',
+            ignoreRestSiblings: true,
+          },
+        ],
+        'no-useless-constructor': 'off',
+        '@typescript-eslint/no-useless-constructor': 'warn',
+        '@typescript-eslint/array-type': ['error', { default: 'array-simple' }],
+      },
+    },
+  ],
 };